Update delta-cache flow to accomodate healing process
Two things being done here:

1. A bookkeeping trick:

  We cache the net result of the deltas into a separate graph (next
  to the files, which are meant for external consumption). This
  will ease the healing process as we can query the difference between
  this graph and other graphs ('ground truth') to check what is missing during healing
  process.

  A side effect which could be potentially beneficial: from this
  graph, a big delta-file could be created, which could be used by a
  newly setup consumer for an initial sync.

2. We perform a check before producing delta files (and CACHE_GRAPH)
to see whether a specic part of the healing process, i.e. the
diffing. During the diffing of the CACHE_GRAPH and whatever we
consider 'ground truth' we want as few as much moving parts.
So no updates of the CACHE_GRAPH as long as the diffing is going.

Note 1: it might be that the blocking mechanism/conditions change of
time. Working towards a tanglible case, and improve from there on.

Note 2: it is a highly async process, I already can imagine cases
where the blocking can go wrong during a specific run. E.g. While
writing to CACHE_GRAPH, the healing/diff task kicks in, diffing from
moving targets.
A potential mitigation path:
 - eventual consistency, achieved in a second healing run (e.g. the next day)
 - a two step lock (so healing process waits for the producing job to
 finish).

To be continued...